created game_logic.rb file and populated it with only logic files
