+++ bin/main.rb
@@ -1,126 +1,68 @@
 #!/usr/bin/env ruby
-class Game
-  def initialize(player1, player2)
-    @player1 = player1
-    @player2 = player2
-    @turn_count = 0
-    @field = ['',
-              '1', '2', '3',
-              '4', '5', '6',
-              '7', '8', '9'] # We ignore index 0 for simplicity
-  end
+require_relative '../lib/game_logic.rb'
 
-  def display_board
-    @field.each_index do |i|
-      if (i % 3).zero?
-        p @field[(i - 2)..i] unless i.zero?
-      end
-    end
-    puts
-  end
-
-  def get_move(player)
-    display_board
-
-    print "(#{player.marker}) Enter choice number, #{player.name}: "
-    move = gets.chomp.to_i
-    player.moves_history << move
-    player_history_sorted = player.moves_history.sort.join
-
-    if @field[move] != 'X' && @field[move] != 'O' && (1..9).include?(move)
-      @field[move] = player.marker
-    else
-      puts 'Invalid Move'
-      get_move(player)
-    end
-
-    check_win(player, player_history_sorted)
-  end
-
-  def game_over
-    print 'Play again? (Y/N): '
-    input = gets.chomp.to_s.downcase
-
-    if input == 'y'
-      Game.start
-    elsif input == 'n'
-      puts 'Thanks for trying!'
-      exit
-    else
-      puts "select 'Y or  'N'."
-      game_over
+def display_board
+  @field.each_index do |i|
+    if (i % 3).zero?
+      p @field[(i - 2)..i] unless i.zero?
     end
   end
+  puts
+end
 
-  def check_win(player, player_history_sorted)
-    WIN.each do |r1|
-      counter = 0
+def get_move(player)
+  display_board
 
-      r1.each do |r2|
-        counter += 1 if player_history_sorted.include?(r2.to_s)
+  print "(#{player.marker}) Enter choice number, #{player.name}: "
+  move = gets.chomp.to_i
+  player.moves_history << move
+  player_history_sorted = player.moves_history.sort.join
 
-        next unless counter == 3
-
-        display_board
-        puts "Wow, #{player.name}! It's a win!"
-        game_over
-      end
-    end
-
-    if @turn_count >= 9
-      display_board
-      puts "It's a tie.Try again?"
-      game_over
-    end
-
-    turn
-  end
-
-  def turn
-    @turn_count += 1
-
-    player = @turn_count.odd? ? @player1 : @player2
-
+  if @field[move] != 'X' && @field[move] != 'O' && (1..9).include?(move)
+    @field[move] = player.marker
+  else
+    puts 'Invalid Move'
     get_move(player)
   end
 
-  def self.start
-    print "Lets start\n"
-    p1_name = nil
-    loop do
-      print 'Player one: '
-      p1_name = gets.chomp
-      break unless p1_name.empty?
+  check_win(player, player_history_sorted)
+end
 
-      puts 'Please Enter at least one character'
-    end
+def game_over
+  print 'Play again? (Y/N): '
+  input = gets.chomp.to_s.downcase
 
-    p2_name = nil
-    loop do
-      print 'Player two: '
-      p2_name = gets.chomp
-      break unless p2_name.empty?
-    end
-
-    p1 = Player.new(p1_name, 'X')
-    p2 = Player.new(p2_name, 'O')
-    @game = Game.new(p1, p2)
-
-    @game.turn
+  if input == 'y'
+    Game.start
+  elsif input == 'n'
+    puts 'Thanks for trying!'
+    exit
+  else
+    puts "select 'Y or  'N'."
+    game_over
   end
 end
 
-class Player
-  attr_accessor :name, :moves_history
-  attr_reader :marker
+def check_win(player, player_history_sorted)
+  WIN.each do |r1|
+    counter = 0
 
-  def initialize(name, marker)
-    @name = name
-    @marker = marker
-    @moves_history = []
+    r1.each do |r2|
+      counter += 1 if player_history_sorted.include?(r2.to_s)
+
+      next unless counter == 3
+
+      display_board
+      puts "Wow, #{player.name}! It's a win!"
+      game_over
+    end
   end
+
+  if @turn_count >= 9
+    display_board
+    puts "It's a tie.Try again?"
+    game_over
+  end
+
+  turn
 end
-
-WIN = [[1, 2, 3], [4, 5, 6], [7, 8, 9], [1, 4, 7], [2, 5, 8], [3, 6, 9], [1, 5, 9], [3, 5, 7]].freeze
-
-Game.start
